# openapi.yaml
openapi: 3.1.0
info:
  title: Restricted Diet Cookbook AI API
  version: 1.0.0
  description: |
    API for generating personalized recipes.
    
    **Principles:**
    - Contract-First (OpenAPI is source of truth).
    - Cookie-based Auth (httpOnly).
    - Cursor Pagination for lists.
    - RFC 7807 Problems.
servers:
  - url: http://localhost:8000
    description: Local Development
  - url: https://cookbook-api.onrender.com
    description: Production
components:
  schemas:
...
    Uuid:
      type: string
      format: uuid
      example: "123e4567-e89b-12d3-a456-426614174000"
    Timestamp:
      type: string
      format: date-time
      example: "2023-10-27T10:00:00Z"

    # --- Errors ---
...
    DietaryRestriction:
      type: string
      enum: [vegan, vegetarian, gluten_free, dairy_free, nut_free, keto, paleo]
    Recipe:
      type: object
      required: [id, title, ingredients, instructions, dietaryTags, created_at]
      properties:
        id:
          $ref: '#/components/schemas/Uuid'
        title:
          type: string
        description:
          type: string
        ingredients:
          type: array
          items:
            type: string
        instructions:
          type: array
          items:
            type: string
        dietaryTags:
          type: array
          items:
            $ref: '#/components/schemas/DietaryRestriction'
        prepTimeMinutes:
          type: integer
        cookTimeMinutes:
          type: integer
        servings:
          type: integer
        difficulty:
          type: string
          enum: [Easy, Medium, Hard]
        calories:
          type: integer
        created_at:
          $ref: '#/components/schemas/Timestamp'
        imageUrl:
          type: string
          format: uri
          nullable: true
        userId:
          $ref: '#/components/schemas/Uuid'
    RecipeList:
      type: object
      required: [data, hasMore]
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/Recipe'
        nextCursor:
          type: string
          nullable: true
          description: "Base64 encoded cursor for next page"
        hasMore:
          type: boolean
    RecipeCreate:
      type: object
      required: [title, ingredients, instruction_text, dietary_tags]
      properties:
        title:
          type: string
        description:
          type: string
        ingredients:
          type: array
          items:
            type: string
        instruction_text:
          type: string
          description: "Raw text or steps"
        dietary_tags:
          type: array
          items:
            type: string
...
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: session_id
    csrfHeader:
      type: apiKey
      in: header
      name: X-CSRF-Token
security:
  - cookieAuth: []
  - csrfHeader: [] # Applied globally, but safe methods ignore it via checks
paths:
...
  /recipes:
    get:
      summary: List Recipes
      parameters:
        - name: cursor
          in: query
          schema:
            type: string
          description: Pagination cursor
        - name: q
          in: query
          schema:
            type: string
          description: Search query text
        - name: limit
          in: query
          schema:
            type: integer
            default: 10
            maximum: 100
        - name: restrictions
          in: query
          style: form
          explode: true
          schema:
            type: array
            items:
              $ref: '#/components/schemas/DietaryRestriction'
      responses:
        '200':
          description: List of recipes
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RecipeList'
    
    post:
      summary: Create Recipe Manually
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RecipeCreate'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Recipe'
  /recipes/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          $ref: '#/components/schemas/Uuid'
    get:
      summary: Get Recipe Details
      responses:
        '200':
          description: Recipe details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Recipe'
        '404':
          description: Not found

    patch:
      summary: Update Recipe
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RecipeCreate' 
              # In reality, verify partial updates, reusing Create schema for brevity in MVP
      responses:
        '200':
          description: Updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Recipe'

    delete:
      summary: Delete Recipe
      responses:
        '204':
          description: Deleted
        '403':
          description: Forbidden
        '404':
          description: Not found
...
  /ai/recipe:
    post:
      summary: Generate Recipe with AI
      operationId: generateRecipe
      tags:
        - ai
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                ingredients:
                  type: array
                  items:
                    type: string
                restrictions:
                  type: array
                  items:
                    type: string
                uploadId:
                  $ref: '#/components/schemas/Uuid'
      responses:
        '200':
          description: Recipe generated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Recipe'
        '401':
          description: Unauthorized
        '402':
          description: Cost limit exceeded